% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CROModelFunctions.R
\name{plotEmissions}
\alias{plotEmissions}
\title{Plot emission profile obtained from CRO-model}
\usage{
plotEmissions(results)
}
\arguments{
\item{results}{A list object as returned by \code{\link{runCROModel}}}
}
\description{
Plots the emission profile of the scenario computed by the CRO-model.
}
