% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CROModelFunctions.R
\name{plotMACCs}
\alias{plotMACCs}
\title{Plot marginal abatement cost curves (MACCs)}
\usage{
plotMACCs(
  results,
  scale = "linear",
  x_min = 0,
  x_max = "max",
  y_min = 0,
  y_max = "max"
)
}
\arguments{
\item{results}{A list object as returned by \code{\link{runCROModel}}}

\item{scale}{y-axis scale (costs). Use "log" for logarithmic scale.}

\item{x_min}{Min value on the x-axis (abatement rate).}

\item{x_max}{Max value on the x-axis (abatement rate).}

\item{y_min}{Min value on the y-axis (costs).}

\item{y_max}{Max value on the y-axis (costs).}
}
\description{
Plots marginal abatement cost curves (MACCs) from the given scenario specification which is used by the CRO-model.
}
