% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CROModelFunctions.R
\name{runCROModel}
\alias{runCROModel}
\title{Execute the CRO-model}
\usage{
runCROModel(model, ssp, rcp, r_d, dacs_cost, dacs_cap)
}
\arguments{
\item{model}{Character string to define Integrated Assessment Model (IAM).}

\item{ssp}{Numerical value to define Shared Socioeconomic Pathway (SSP)}

\item{rcp}{Numerical value to define climate target as Representative Concentration Pathway (RCP).}

\item{r_d}{Numerical value to define interest rate on carbon debt.}

\item{dacs_cost}{Character string to define direct air caputure and storage (DACS) marginal abatement costs (MAC).}

\item{dacs_cap}{Character string  to define capacity limit of direct air caputure and storage (DACS).}
}
\value{
A list-object containing detailed results, the scenario specification and model parameters.

\itemize{
\item $data contains annual results,
\item $scenario contains the defined scenario specification,
\item $parameters contains the scenario's parameters (e.g. specification of marginal cost curves, etc.)

}
}
\description{
Executes the CRO-model for a scenario specification and plots resultant emission paths.
}
\details{
The correct scenario specification can be obtained using \code{\link{showScenarios}}
}
\examples{
\notrun{
 setwd("~/gamsdir/CROmodel/")
 results <- runCROModel(model="MESSAGE-GLOBIOM", ssp=2, rcp=19, r_d=0.01, dacs_cost="HiCost", dacs_cap="LoCap")
}


}
