% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CROModelFunctions.R
\name{showScenarios}
\alias{showScenarios}
\title{Show scenario specifications serving as input to the CRO-model}
\usage{
showScenarios(
  model = "all",
  ssp = "all",
  rcp = "all",
  r_d = "all",
  dacs_cost = "all",
  dacs_cap = "all"
)
}
\arguments{
\item{model}{Character string  to restrict selection to one Integrated Assessment Model (IAM). By default all models are shown.}

\item{ssp}{Numerical value to restrict selection to one Shared Socioeconomic Pathway (SSP). By default all SSPs are shown.}

\item{rcp}{Numerical value to restrict selection to one Representative Concentration Pathway (RCP). By default all RCPs are shown.}

\item{r_d}{Numerical value to restrict selection to one interest rate on carbon debt. By default all rates are shown.}

\item{dacs_cost}{Character string  to restrict selection to low, medium or high direct air caputure and storage (DACS) marginal abatement costs (MAC). By default all MAC-curves are shown.}

\item{dacs_cap}{Character string  to restrict selection to low or high capacity direct air caputure and storage (DACS). By default all MAC-curves are shown.}
}
\value{
A data.frame containing all scenario specifications compatible with the given inputs.
}
\description{
Shows all possible scenarios that can be computed using the CRO-model.
Use function parameters to filter for subsets of possible scenarios.
}
